+++ subscription/mongodb/common/blocking/competing-consumer-strategy/src/main/java/org/occurrent/subscription/mongodb/blocking/ccs/internal/MongoLeaseCompetingConsumerStrategySupport.java
@@ -59,6 +59,7 @@ public class MongoLeaseCompetingConsumerStrategySupport {
         this.clock = clock;
         this.leaseTime = leaseTime;
         this.scheduledRefresh = ScheduledRefresh.auto();
+        this.running = true;
         this.competingConsumerListeners = Collections.newSetFromMap(new ConcurrentHashMap<>());
         this.competingConsumers = new ConcurrentHashMap<>();