+++ changelog.md
@@ -3,6 +3,7 @@
 * Fixed issue in `MongoListenerLockService` (used by competing subscription models) in which two subscribers could "race" to catch the lease one more time unnecessarily.
 * Upgraded spring-boot from 3.4.1 to 3.4.2
 * Upgraded spring-data-mongodb from 4.4.2 to 4.4.3
+* Upgraded mongodb-driver-sync from 5.3.0 to 5.3.1
 
 ### 0.19.8 (2025-01-17)
 * Converted `org.occurrent.subscription.OccurrentSubscriptionFilter` from a Java class to a record. This means that the `public final` filter instance field is now a record property. So if you ever used `occurrentSubscriptionFilter.filter` to access the underlying filter, you now need to do `occurrentSubscriptionFilter.filter()` instead.