+++ changelog.md
@@ -7,6 +7,7 @@
 * Upgraded jobrunr from 7.3.2 to 7.4.0
 * Upgraded kotlin from 2.1.0 to 2.1.10
 * Upgraded project reactor from 3.7.2 to 3.7.3
+* Upgraded xstream from 1.4.20 to 1.4.21
 
 ### 0.19.8 (2025-01-17)
 * Converted `org.occurrent.subscription.OccurrentSubscriptionFilter` from a Java class to a record. This means that the `public final` filter instance field is now a record property. So if you ever used `occurrentSubscriptionFilter.filter` to access the underlying filter, you now need to do `occurrentSubscriptionFilter.filter()` instead.