+++ changelog.md
@@ -1,4 +1,4 @@
-### Next version
+### 0.19.9 (2025-02-14)
 * Added toString() to subscription models for better debug output
 * Fixed issue in `MongoListenerLockService` (used by competing subscription models) in which two subscribers could "race" to catch the lease one more time unnecessarily.
 * Upgraded spring-boot from 3.4.1 to 3.4.2